+++ src/base/b-virtual-scroll/test/runners/functional/state.js
@@ -24,14 +24,6 @@ module.exports = (page) => {
 		node,
 		container;
 
-	beforeEach(async () => {
-		await h.utils.reloadAndWaitForIdle(page);
-
-		component = await h.component.waitForComponent(page, '#target');
-		node = await h.dom.waitForEl(page, '#target');
-		container = await h.dom.waitForRef(node, 'container');
-	});
-
 	const
 		getArray = (offset = 0, length = 12) => ({data: Array.from(Array(length), (v, i) => ({i: i + offset}))}),
 		firstChunkExpected = getArray(),
@@ -66,16 +58,51 @@ module.exports = (page) => {
 	const setProps = (requestProps = {}) => component.evaluate((ctx, requestProps) => {
 		ctx.dataProvider = 'demo.Pagination';
 		ctx.chunkSize = 10;
-		ctx.request = {get: {chunkSize: 12, id: 'uniq', ...requestProps}};
+		ctx.request = {get: {chunkSize: 12, id: Math.random(), ...requestProps}};
 		ctx.componentConverter = (v) => ({data: v.data});
 	}, requestProps);
+
+	beforeEach(async () => {
+		await h.utils.reloadAndWaitForIdle(page);
+		await h.component.waitForComponent(page, '#root-component');
+
+		await page.evaluate(() => {
+			globalThis.removeCreatedComponents();
+
+			const baseAttrs = {
+				theme: 'demo',
+				option: 'section',
+				optionProps: ({current}) => ({'data-index': current.i})
+			};
+
+			const scheme = [
+				{
+					attrs: {
+						...baseAttrs,
+						id: 'target'
+					}
+				}
+			];
+
+			globalThis.renderComponents('b-virtual-scroll', scheme);
+		});
+
+		await h.bom.waitForIdleCallback(page);
+		await h.component.waitForComponentStatus(page, '.b-virtual-scroll', 'ready');
+
+		component = await h.component.waitForComponent(page, '#target');
+		node = await h.dom.waitForEl(page, '#target');
+		container = await h.dom.waitForRef(node, 'container');
+	});
 
 	describe('b-virtual-scroll getCurrentDataState', () => {
 		describe('returns the correct value', () => {
 			it('if there is no `dataProvider`', async () => {
-				const expected = getExpected({currentPage: 0, nextPage: 1});
+				const
+					expected = getExpected(),
+					current = await getCurrentComponentState();
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				expect(current).toEqual(expected);
 			});
 
 			it('after loading the first chunk', async () => {
@@ -94,7 +121,8 @@ module.exports = (page) => {
 				await setProps();
 				await h.dom.waitForEl(container, 'section');
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				const current = await getCurrentComponentState();
+				expect(current).toEqual(expected);
 			});
 
 			it('after loading the second chunk', async () => {
@@ -116,7 +144,8 @@ module.exports = (page) => {
 				await h.scroll.scrollToBottom(page);
 				await h.dom.waitForEl(container, 'section:nth-child(11)');
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				const current = await getCurrentComponentState();
+				expect(current).toEqual(expected);
 			});
 
 			it('after re-initialization and without `dataProvider`', async () => {
@@ -132,9 +161,10 @@ module.exports = (page) => {
 				});
 
 				await h.dom.waitForEl(container, 'section', {to: 'unmount'});
-				await h.bom.waitForIdleCallback(page, {sleepAfterIdles: 200});
+				await h.bom.waitForIdleCallback(page, {sleepAfterIdles: 1000});
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				const current = await getCurrentComponentState();
+				expect(current).toEqual(expected);
 			});
 
 			it('after re-initialization and with `dataProvider`', async () => {
@@ -153,11 +183,12 @@ module.exports = (page) => {
 				await setProps();
 				await h.dom.waitForEl(container, 'section');
 
-				await setProps({id: 'new-id'});
+				await setProps({id: Math.random()});
 				await h.dom.waitForEl(container, 'section', {to: 'unmount'});
 				await h.dom.waitForEl(container, 'section', {to: 'mount'});
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				const current = await getCurrentComponentState();
+				expect(current).toEqual(expected);
 			});
 
 			it('if for the full loading it was necessary to go several times to `dataProvider`', async () => {
@@ -176,7 +207,8 @@ module.exports = (page) => {
 				await setProps({chunkSize: 4});
 				await h.dom.waitForEl(container, 'section');
 
-				expect(await getCurrentComponentState()).toEqual(expected);
+				const current = await getCurrentComponentState();
+				expect(current).toEqual(expected);
 			});
 		});
 	});
@@ -184,21 +216,25 @@ module.exports = (page) => {
 	describe('b-virtual-scroll getDataStateSnapshot', () => {
 		describe('returns the correct value', () => {
 			it('with `chunkRequest` and `chunkRender`', async () => {
-				const expected = getExpected();
+				const
+					expected = getExpected(),
+					current = await component.evaluate((ctx) => ctx.getDataStateSnapshot({
+						items: undefined,
+						itemsTillBottom: undefined
+					}));
 
-				expect(await component.evaluate((ctx) => ctx.getDataStateSnapshot({
-					items: undefined,
-					itemsTillBottom: undefined
-				}))).toEqual(expected);
+				expect(current).toEqual(expected);
 			});
 
 			it('with `chunkRequest`', async () => {
-				const expected = getExpected();
+				const
+					expected = getExpected(),
+					current = await component.evaluate((ctx) => ctx.getDataStateSnapshot({
+						items: undefined,
+						itemsTillBottom: undefined
+					}, ctx.chunkRequest));
 
-				expect(await component.evaluate((ctx) => ctx.getDataStateSnapshot({
-					items: undefined,
-					itemsTillBottom: undefined
-				}, ctx.chunkRequest))).toEqual(expected);
+				expect(current).toEqual(expected);
 			});
 
 			it('with override params, `chunkRequest` and `chunkRender`', async () => {
@@ -207,10 +243,12 @@ module.exports = (page) => {
 					nextPage: 2
 				});
 
-				expect(await component.evaluate((ctx) => ctx.getDataStateSnapshot({
+				const current = await component.evaluate((ctx) => ctx.getDataStateSnapshot({
 					items: undefined,
 					itemsTillBottom: undefined
-				}, ctx.chunkRequest, ctx.chunkRender))).toEqual(expected);
+				}, ctx.chunkRequest, ctx.chunkRender));
+
+				expect(current).toEqual(expected);
 			});
 		});
 	});