+++ src/core/prelude/test-env/CHANGELOG.md
@@ -9,8 +9,8 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.0.0-rc.61 ()
+## v3.0.0-rc.60 (2020-09-01)
 
-#### :nail_care: Polish
+#### :rocket: New Feature
 
-* Added documentation
+* The module has been created