+++ src/icons/b-icon/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.0.0-rc.60 ()
+## v3.0.0-rc.61 ()
 
 #### :house: Internal